Fixed project.toml and added a unittest for PyPI
