+++ pyproject.toml
@@ -77,17 +77,18 @@ Repository = "https://github.com/markrussinovich/refchecker"
 "Bug Tracker" = "https://github.com/markrussinovich/refchecker/issues"
 
 [project.scripts]
-academic-refchecker = "core.refchecker:main"
+academic-refchecker = "refchecker.core.refchecker:main"
 
 [tool.setuptools.packages.find]
-where = ["src"]
+# Look in root directory for packages
+where = ["."]
+# Only include the refchecker package and subpackages
+include = ["refchecker*"]
 
-[tool.setuptools]
-py-modules = ["__version__"]
 
 [tool.setuptools.dynamic]
-version = {attr = "__version__.__version__"}
+version = {attr = "refchecker.__version__.__version__"}
 
 [tool.setuptools.package-data]
 "*" = ["*.txt", "*.md", "*.conf", "*.py"]
-"scripts" = ["*.py"]
+"refchecker.scripts" = ["*.py"]
