+++ refchecker.py
@@ -9,19 +9,18 @@ with the proper package structure.
 import sys
 import os
 
-# Add the src directory to Python path
-sys.path.insert(0, os.path.join(os.path.dirname(__file__), 'src'))
+# Add the package directory to Python path
+sys.path.insert(0, os.path.dirname(__file__))
 
 # Import all modules to ensure they're available
-import checkers.semantic_scholar
-import checkers.local_semantic_scholar
-import checkers.openalex
-import checkers.crossref
-import checkers.enhanced_hybrid_checker
-import utils.text_utils
-import utils.author_utils
-
-from core.refchecker import main
+import refchecker.checkers.semantic_scholar
+import refchecker.checkers.local_semantic_scholar
+import refchecker.checkers.openalex
+import refchecker.checkers.crossref
+import refchecker.checkers.enhanced_hybrid_checker
+import refchecker.utils.text_utils
+import refchecker.utils.author_utils
+from refchecker.core.refchecker import main
 
 if __name__ == "__main__":
     main()
