+++ refchecker/checkers/crossref.py
@@ -1,25 +1,23 @@
 #!/usr/bin/env python3
 """
-OpenAlex API Client for Reference Verification
+CrossRef API Client for Reference Verification
 
-This module provides functionality to verify non-arXiv references using the OpenAlex API.
-It can check if a reference's metadata (authors, year, title) matches what's in the OpenAlex database.
-
-OpenAlex is a comprehensive database of scholarly papers, authors, venues, institutions, and concepts.
-It provides free access to over 240 million research outputs with excellent coverage across all disciplines.
+This module provides functionality to verify references using the CrossRef API.
+CrossRef maintains metadata for over 165 million research outputs from 20,000+ members
+and is particularly strong for publisher-registered content with DOIs.
 
 Usage:
-    from openalex import OpenAlexReferenceChecker
+    from crossref import CrossRefReferenceChecker
     
     # Initialize the checker
-    checker = OpenAlexReferenceChecker(email="your@email.com")  # Email for polite pool
+    checker = CrossRefReferenceChecker(email="your@email.com")  # Email for polite pool
     
     # Verify a reference
     reference = {
         'title': 'Title of the paper',
         'authors': ['Author 1', 'Author 2'],
         'year': 2020,
-        'url': 'https://example.com/paper',
+        'doi': '10.1000/xyz123',
         'raw_text': 'Full citation text'
     }
     
@@ -32,9 +30,9 @@ import logging
 import re
 from typing import Dict, List, Tuple, Optional, Any, Union
 from urllib.parse import quote_plus
-from utils.text_utils import normalize_text, clean_title_basic, find_best_match, is_name_match, compare_authors, clean_title_for_search
-from utils.error_utils import format_year_mismatch, format_doi_mismatch
-from config.settings import get_config
+from refchecker.utils.text_utils import normalize_text, clean_title_basic, find_best_match, is_name_match, compare_authors, clean_title_for_search
+from refchecker.utils.error_utils import format_year_mismatch, format_doi_mismatch
+from refchecker.config.settings import get_config
 
 # Set up logging
 logger = logging.getLogger(__name__)
@@ -43,19 +41,19 @@ logger = logging.getLogger(__name__)
 config = get_config()
 SIMILARITY_THRESHOLD = config["text_processing"]["similarity_threshold"]
 
-class OpenAlexReferenceChecker:
+class CrossRefReferenceChecker:
     """
-    A class to verify non-arXiv references using the OpenAlex API
+    A class to verify references using the CrossRef API
     """
     
     def __init__(self, email: Optional[str] = None):
         """
-        Initialize the OpenAlex API client
+        Initialize the CrossRef API client
         
         Args:
             email: Optional email for polite pool access (better performance)
         """
-        self.base_url = "https://api.openalex.org"
+        self.base_url = "https://api.crossref.org"
         self.headers = {
             "Accept": "application/json",
             "User-Agent": "RefChecker/1.0.0 (https://github.com/markrussinovich/refchecker)"
@@ -63,12 +61,12 @@ class OpenAlexReferenceChecker:
         
         # Add email to headers for polite pool access
         if email:
-            self.headers["User-Agent"] += f" mailto:{email}"
+            self.headers["User-Agent"] += f"; mailto:{email}"
         
-        # Rate limiting parameters - OpenAlex allows 10 requests per second
-        self.request_delay = 0.1  # 100ms between requests (10 req/sec)
+        # Rate limiting parameters - CrossRef has variable rate limits
+        self.request_delay = 0.05  # 50ms between requests (20 req/sec conservative)
         self.max_retries = 3
-        self.backoff_factor = 2  # Exponential backoff factor
+        self.backoff_factor = 2
     
     def search_works(self, query: str, year: Optional[int] = None, limit: int = 5) -> List[Dict[str, Any]]:
         """
@@ -84,26 +82,34 @@ class OpenAlexReferenceChecker:
         """
         endpoint = f"{self.base_url}/works"
         
-        # Build query parameters - OpenAlex uses a flexible search syntax
         params = {
-            "search": query,
-            "per_page": min(limit, 25),  # OpenAlex max per page is 200, but we limit for performance
-            "select": "id,doi,title,display_name,publication_year,authorships,type,open_access,primary_location,locations,referenced_works,ids"
+            "query": query,
+            "rows": min(limit, 20),  # Limit for performance
+            "select": "DOI,title,author,published,publisher,container-title,type,URL,link,abstract,subject"
         }
         
         # Add year filter if provided
         if year:
-            params["filter"] = f"publication_year:{year}"
+            params["filter"] = f"from-pub-date:{year},until-pub-date:{year}"
         
         # Make the request with retries and backoff
         for attempt in range(self.max_retries):
-            try:            
+            try:
+                # Add delay to respect rate limits
+                time.sleep(self.request_delay)
+                
                 response = requests.get(endpoint, headers=self.headers, params=params, timeout=30)
                 
-                # Check for rate limiting (OpenAlex returns 429 for rate limits)
+                # Check for rate limiting
                 if response.status_code == 429:
-                    wait_time = self.request_delay * (self.backoff_factor ** attempt) + 1
-                    logger.debug(f"OpenAlex rate limit exceeded. Retrying in {wait_time:.2f} seconds...")
+                    # Check if rate limit info is in headers
+                    retry_after = response.headers.get('Retry-After')
+                    if retry_after:
+                        wait_time = int(retry_after) + 1
+                    else:
+                        wait_time = self.request_delay * (self.backoff_factor ** attempt) + 1
+                    
+                    logger.debug(f"CrossRef rate limit exceeded. Retrying in {wait_time} seconds...")
                     time.sleep(wait_time)
                     continue
                 
@@ -112,18 +118,18 @@ class OpenAlexReferenceChecker:
                 
                 # Parse the response
                 data = response.json()
-                results = data.get('results', [])
+                results = data.get('message', {}).get('items', [])
                 
-                logger.debug(f"OpenAlex search returned {len(results)} results for query: {query[:50]}...")
+                logger.debug(f"CrossRef search returned {len(results)} results for query: {query[:50]}...")
                 return results
                 
             except requests.exceptions.RequestException as e:
                 wait_time = self.request_delay * (self.backoff_factor ** attempt) + 1
-                logger.debug(f"OpenAlex request failed: {str(e)}. Retrying in {wait_time:.2f} seconds...")
+                logger.debug(f"CrossRef request failed: {str(e)}. Retrying in {wait_time:.2f} seconds...")
                 time.sleep(wait_time)
         
         # If we get here, all retries failed
-        logger.debug(f"Failed to search OpenAlex after {self.max_retries} attempts")
+        logger.warning(f"Failed to search CrossRef after {self.max_retries} attempts")
         return []
     
     def get_work_by_doi(self, doi: str) -> Optional[Dict[str, Any]]:
@@ -145,11 +151,11 @@ class OpenAlexReferenceChecker:
         if doi.startswith('http://doi.org/'):
             clean_doi = doi[15:]
         
-        endpoint = f"{self.base_url}/works/doi:{clean_doi}"
+        endpoint = f"{self.base_url}/works/{clean_doi}"
         
-        params = {
-            "select": "id,doi,title,display_name,publication_year,authorships,type,open_access,primary_location,locations,referenced_works,ids"
-        }
+        # Note: The individual DOI endpoint does not support the 'select' parameter
+        # It returns all fields by default, which is what we want
+        params = {}
         
         # Make the request with retries and backoff
         for attempt in range(self.max_retries):
@@ -161,31 +167,37 @@ class OpenAlexReferenceChecker:
                 
                 # Check for rate limiting
                 if response.status_code == 429:
-                    wait_time = self.request_delay * (self.backoff_factor ** attempt) + 1
-                    logger.debug(f"OpenAlex rate limit exceeded. Retrying in {wait_time:.2f} seconds...")
+                    retry_after = response.headers.get('Retry-After')
+                    if retry_after:
+                        wait_time = int(retry_after) + 1
+                    else:
+                        wait_time = self.request_delay * (self.backoff_factor ** attempt) + 1
+                    
+                    logger.debug(f"CrossRef rate limit exceeded. Retrying in {wait_time} seconds...")
                     time.sleep(wait_time)
                     continue
                 
                 # If not found, return None
                 if response.status_code == 404:
-                    logger.debug(f"Work with DOI {doi} not found in OpenAlex")
+                    logger.debug(f"Work with DOI {doi} not found in CrossRef")
                     return None
                 
                 # Check for other errors
                 response.raise_for_status()
                 
                 # Parse the response
-                work_data = response.json()
-                logger.debug(f"Found work by DOI in OpenAlex: {doi}")
+                data = response.json()
+                work_data = data.get('message', {})
+                logger.debug(f"Found work by DOI in CrossRef: {doi}")
                 return work_data
                 
             except requests.exceptions.RequestException as e:
                 wait_time = self.request_delay * (self.backoff_factor ** attempt) + 1
-                logger.debug(f"OpenAlex request failed: {str(e)}. Retrying in {wait_time:.2f} seconds...")
+                logger.warning(f"CrossRef request failed: {str(e)}. Retrying in {wait_time:.2f} seconds...")
                 time.sleep(wait_time)
         
         # If we get here, all retries failed
-        logger.error(f"Failed to get work by DOI from OpenAlex after {self.max_retries} attempts")
+        logger.error(f"Failed to get work by DOI from CrossRef after {self.max_retries} attempts")
         return None
     
     def extract_doi_from_url(self, url: str) -> Optional[str]:
@@ -238,24 +250,31 @@ class OpenAlexReferenceChecker:
         # Use common normalization function
         return normalize_text(name)
     
-    def compare_authors(self, cited_authors: List[str], openalex_authors: List[Dict[str, Any]]) -> Tuple[bool, str]:
+    def compare_authors(self, cited_authors: List[str], crossref_authors: List[Dict[str, Any]]) -> Tuple[bool, str]:
         """
         Compare author lists to check if they match (delegates to shared utility)
         
         Args:
             cited_authors: List of author names as cited
-            openalex_authors: List of authorship data from OpenAlex
+            crossref_authors: List of author data from CrossRef
             
         Returns:
             Tuple of (match_result, error_message)
         """
-        # Extract author names from OpenAlex data for the shared utility
+        # Extract author names from CrossRef data for the shared utility
         author_dicts = []
-        for authorship in openalex_authors:
-            author = authorship.get('author', {})
-            display_name = author.get('display_name', '')
-            if display_name:
-                author_dicts.append({'name': display_name})
+        for author in crossref_authors:
+            # CrossRef author format: {"given": "First", "family": "Last", "name": "Full Name"}
+            name = None
+            if 'name' in author:
+                name = author['name']
+            elif 'given' in author and 'family' in author:
+                name = f"{author['given']} {author['family']}"
+            elif 'family' in author:
+                name = author['family']
+            
+            if name:
+                author_dicts.append({'name': name})
         
         return compare_authors(cited_authors, author_dicts)
     
@@ -299,83 +318,78 @@ class OpenAlexReferenceChecker:
         
         return True
     
-    def extract_url_from_work(self, work_data: Dict[str, Any]) -> Optional[str]:
+    def extract_year_from_published(self, published: Dict[str, List[int]]) -> Optional[int]:
         """
-        Extract the best URL from OpenAlex work data
+        Extract year from CrossRef published date
         
         Args:
-            work_data: Work data from OpenAlex
+            published: Published date object from CrossRef
+            
+        Returns:
+            Publication year or None
+        """
+        if not published:
+            return None
+        
+        # CrossRef date format: {"date-parts": [[2017, 6, 12]]}
+        date_parts = published.get('date-parts', [])
+        if date_parts and len(date_parts) > 0 and len(date_parts[0]) > 0:
+            return date_parts[0][0]  # First element is the year
+        
+        return None
+    
+    def extract_url_from_work(self, work_data: Dict[str, Any]) -> Optional[str]:
+        """
+        Extract the best URL from CrossRef work data
+        
+        Args:
+            work_data: Work data from CrossRef
             
         Returns:
             Best available URL or None
         """
-        # Priority order: Open access PDF, primary location, DOI
+        # Priority order: Direct URL, DOI URL, Link URLs
         
-        # Check for open access PDF
-        open_access = work_data.get('open_access', {})
-        if open_access.get('is_oa') and open_access.get('oa_url'):
-            logger.debug(f"Found open access URL: {open_access['oa_url']}")
-            return open_access['oa_url']
+        # Check for direct URL
+        if work_data.get('URL'):
+            logger.debug(f"Found direct URL: {work_data['URL']}")
+            return work_data['URL']
         
-        # Check primary location
-        primary_location = work_data.get('primary_location', {})
-        if primary_location:
-            # Try landing page URL first
-            if primary_location.get('landing_page_url'):
-                logger.debug(f"Found primary location URL: {primary_location['landing_page_url']}")
-                return primary_location['landing_page_url']
-            
-            # Try PDF URL
-            if primary_location.get('pdf_url'):
-                logger.debug(f"Found primary location PDF: {primary_location['pdf_url']}")
-                return primary_location['pdf_url']
-        
-        # Check other locations for PDFs
-        locations = work_data.get('locations', [])
-        for location in locations:
-            if location.get('pdf_url'):
-                logger.debug(f"Found PDF in location: {location['pdf_url']}")
-                return location['pdf_url']
-            if location.get('landing_page_url'):
-                logger.debug(f"Found landing page in location: {location['landing_page_url']}")
-                return location['landing_page_url']
-        
-        # Fall back to DOI URL
-        doi = work_data.get('doi')
+        # Check for DOI
+        doi = work_data.get('DOI')
         if doi:
             from utils.doi_utils import construct_doi_url
             doi_url = construct_doi_url(doi)
             logger.debug(f"Generated DOI URL: {doi_url}")
             return doi_url
         
-        # Check ids for other identifiers
-        ids = work_data.get('ids', {})
-        if ids.get('doi'):
-            from utils.doi_utils import construct_doi_url
-            doi_url = construct_doi_url(ids['doi'])
-            logger.debug(f"Generated DOI URL from ids: {doi_url}")
-            return doi_url
+        # Check link arrays for URLs
+        links = work_data.get('link', [])
+        for link in links:
+            if isinstance(link, dict) and link.get('URL'):
+                logger.debug(f"Found link URL: {link['URL']}")
+                return link['URL']
         
-        logger.debug("No URL found in OpenAlex work data")
+        logger.debug("No URL found in CrossRef work data")
         return None
     
     def verify_reference(self, reference: Dict[str, Any]) -> Tuple[Optional[Dict[str, Any]], List[Dict[str, Any]], Optional[str]]:
         """
-        Verify a non-arXiv reference using OpenAlex
+        Verify a reference using CrossRef
         
         Args:
             reference: Reference data dictionary
             
         Returns:
             Tuple of (verified_data, errors, url)
-            - verified_data: Work data from OpenAlex or None if not found
+            - verified_data: Work data from CrossRef or None if not found
             - errors: List of error dictionaries
             - url: URL of the work if found, None otherwise
         """
         errors = []
         
         # Extract reference data
-        title = reference.get('title', '') or ''
+        title = reference.get('title', '')
         authors = reference.get('authors', [])
         year = reference.get('year', 0)
         url = reference.get('url', '')
@@ -395,9 +409,9 @@ class OpenAlexReferenceChecker:
             work_data = self.get_work_by_doi(doi)
             
             if work_data:
-                logger.debug(f"Found work by DOI in OpenAlex: {doi}")
+                logger.debug(f"Found work by DOI in CrossRef: {doi}")
             else:
-                logger.debug(f"Could not find work with DOI in OpenAlex: {doi}")
+                logger.debug(f"Could not find work with DOI in CrossRef: {doi}")
         
         # If we couldn't get the work by DOI, try searching by title
         if not work_data and title:
@@ -407,36 +421,52 @@ class OpenAlexReferenceChecker:
             # Search for the work
             search_results = self.search_works(cleaned_title, year)
             
-            if search_results:
-                best_match, best_score = find_best_match(search_results, cleaned_title, year, authors)
+            # Process search results for CrossRef format
+            processed_results = []
+            for result in search_results:
+                # CrossRef title format: ["Title of the Paper"]
+                result_titles = result.get('title', [])
+                if result_titles:
+                    result_title = result_titles[0] if isinstance(result_titles, list) else str(result_titles)
+                    # Create a normalized result for the utility function
+                    processed_result = dict(result)
+                    processed_result['title'] = result_title
+                    processed_result['publication_year'] = self.extract_year_from_published(result.get('published'))
+                    processed_results.append(processed_result)
+            
+            if processed_results:
+                best_match, best_score = find_best_match(processed_results, cleaned_title, year, authors)
                 
                 # Use match if score is good enough
                 if best_match and best_score >= SIMILARITY_THRESHOLD:
                     work_data = best_match
-                    logger.debug(f"Found work by title in OpenAlex with score {best_score:.2f}: {cleaned_title}")
+                    logger.debug(f"Found work by title in CrossRef with score {best_score:.2f}: {cleaned_title}")
                 else:
-                    logger.debug(f"No good title match found in OpenAlex (best score: {best_score:.2f})")
+                    logger.debug(f"No good title match found in CrossRef (best score: {best_score:.2f})")
             else:
-                logger.debug(f"No works found for title in OpenAlex: {cleaned_title}")
+                logger.debug(f"No works found for title in CrossRef: {cleaned_title}")
         
         # If we still couldn't find the work, return no verification
         if not work_data:
-            logger.debug("Could not find matching work in OpenAlex")
+            logger.debug("Could not find matching work in CrossRef")
             return None, [], None
         
         # Verify authors
         if authors:
-            authorships = work_data.get('authorships', [])
-            authors_match, author_error = self.compare_authors(authors, authorships)
+            crossref_authors = work_data.get('author', [])
+            authors_match, author_error = self.compare_authors(authors, crossref_authors)
             
             if not authors_match:
                 # Extract correct author names for error reporting
                 correct_author_names = []
-                for authorship in authorships:
-                    author = authorship.get('author', {})
-                    display_name = author.get('display_name', '')
-                    if display_name:
-                        correct_author_names.append(display_name)
+                for author in crossref_authors:
+                    if 'name' in author:
+                        correct_author_names.append(author['name'])
+                    elif 'given' in author and 'family' in author:
+                        full_name = f"{author['given']} {author['family']}"
+                        correct_author_names.append(full_name)
+                    elif 'family' in author:
+                        correct_author_names.append(author['family'])
                 
                 errors.append({
                     'error_type': 'author',
@@ -445,7 +475,7 @@ class OpenAlexReferenceChecker:
                 })
         
         # Verify year
-        work_year = work_data.get('publication_year')
+        work_year = self.extract_year_from_published(work_data.get('published'))
         if year and work_year and year != work_year:
             errors.append({
                 'warning_type': 'year',
@@ -454,10 +484,7 @@ class OpenAlexReferenceChecker:
             })
         
         # Verify DOI
-        work_doi = work_data.get('doi')
-        if not work_doi and work_data.get('ids', {}).get('doi'):
-            work_doi = work_data['ids']['doi']
-        
+        work_doi = work_data.get('DOI')
         if doi and work_doi:
             # Compare DOIs using the proper comparison function
             from utils.doi_utils import compare_dois
@@ -475,14 +502,14 @@ class OpenAlexReferenceChecker:
 
 if __name__ == "__main__":
     # Example usage
-    checker = OpenAlexReferenceChecker(email="test@example.com")
+    checker = CrossRefReferenceChecker(email="test@example.com")
     
     # Example reference
     reference = {
         'title': 'Attention is All You Need',
         'authors': ['Ashish Vaswani', 'Noam Shazeer'],
         'year': 2017,
-        'url': 'https://example.com/paper',
+        'doi': '10.5555/3295222.3295349',
         'raw_text': 'Vaswani, A., Shazeer, N., Parmar, N., Uszkoreit, J., Jones, L., Gomez, A. N., ... & Polosukhin, I. (2017). Attention is all you need. Advances in neural information processing systems, 30.'
     }
     
@@ -490,7 +517,8 @@ if __name__ == "__main__":
     verified_data, errors, url = checker.verify_reference(reference)
     
     if verified_data:
-        print(f"Found work: {verified_data.get('title') or verified_data.get('display_name')}")
+        print(f"Found work: {verified_data.get('title', ['Unknown'])[0]}")
+        print(f"DOI: {verified_data.get('DOI', 'None')}")
         print(f"URL: {url}")
         
         if errors:
